refactor: Moved JobDescriptionSelector to App
