+++ frontend/src/App.css
@@ -51,12 +51,12 @@
 
 /* Recorder Component Styles */
 .recorder-container {
-  display: flex;
+  /* display: flex;
   flex-direction: column;
   align-items: center;
   max-width: 800px;
   margin: 0 auto;
-  padding: 20px;
+  padding: 20px; */
 }
 
 .controls {
