+++ frontend/src/App.tsx
@@ -1,10 +1,21 @@
+import { useState } from "react";
 import "./App.css";
+import { JobDescriptionSelector } from "./components";
 import Footer from "./components/Footer";
 import Recorder from "./components/Recorder";
+import type { JobDescriptionCategory } from "./types/jobDescriptions";
 
 export default function App() {
+  const [selectedJobDescription, setSelectedJobDescription] = useState<JobDescriptionCategory>('java-developer');
+  const [customJobDescription, setCustomJobDescription] = useState<string | undefined>(undefined);
+
+  const handleJobDescriptionChange = (category: JobDescriptionCategory, customDescription?: string) => {
+    setSelectedJobDescription(category);
+    setCustomJobDescription(customDescription);
+  };
+
   return (
-    <div className="s">
+    <div>
       <header className="app-header fade-in">
         <h1>Interview Response Analyser</h1>
         <p>
@@ -15,7 +26,14 @@ export default function App() {
       </header>
 
       <main className="fade-in delay">
-        <Recorder />
+        <JobDescriptionSelector
+          selectedJobDescription={selectedJobDescription}
+          onJobDescriptionChange={handleJobDescriptionChange}
+        />
+        <Recorder
+          selectedJobDescription={selectedJobDescription}
+          customJobDescription={customJobDescription}
+        />
         <Footer />
       </main>
     </div>
